testpy_net/requirements.txt
requests
tqdm
PyExecJS